requirements.txt
#pyside
#python3-tk
pyttk
gpsbabel
pyserial
#matplotlib
stdConfigParser
#python2-biggles
pyudev
#folium
#pyrtlsdr
progressbar
gmplot
pygame